requirements.txt
requests
python-telegram-bot[rate-limiter]
redis
fake_useragent
asyncio
psycopg2-binary
beautifulsoup4
apscheduler
lxml
tabulate
plotly
kaleido